+++ samples/e2eTestUtils/ElectronPlaywrightTestUtils.ts
