+++ samples/package.json
@@ -20,6 +20,7 @@
         "jest": "^27.1.1",
         "puppeteer": "^13.6.0",
         "ts-jest": "^27.0.5",
-        "wait-on": "^5.2.1"
+        "wait-on": "^5.2.1",
+        "playwright": "^1.21.1"
     }
 }
